Avoid massive DB class
